+++ scripts/actors/pc.js
@@ -17,6 +17,13 @@ function newRating() {
 export function o13pcActorMixin(base) {
 	return class o13pcActor extends inventoryActorMixin(base) {
 		//Updates & Create
+		constructor(...args) {
+			super(...args);
+			
+			this._activerollconfig = null;
+			this._lastrolloptions = null;
+		}
+		
 		async _preCreate(data, options, user) {
 			await super._preCreate(data, options, user);
 
@@ -80,11 +87,15 @@ export function o13pcActorMixin(base) {
 						
 						//updates might mess with relations archetypel links
 						for (let i = 0; i < newRelations.length; i++) {
-							newRelations[i].archetype = newRelations[i].archetype || currentRelations[i]?.archetype || siblingArchetypes[i]?.id;
+							newRelations[i].archetype = siblingArchetypes[i]?.id || newRelations[i].archetype || currentRelations[i]?.archetype;
 						}
 					}
 				}
 			}
+			
+			options.o13 = options.o13 ?? {};
+			
+			options.o13.shouldSyncImage = this.shouldSynchArchetypePortrait;
 			
 			await super._preUpdate(changed, options, user);
 		}
@@ -97,6 +108,10 @@ export function o13pcActorMixin(base) {
 					if (changed.system.hasOwnProperty("archetype")) {
 						await this.updateArchetypeItems();
 						await this.synctoArchetypeBackground();
+						
+						if (options.o13?.shouldSyncImage) {
+							await this.synchArchetypePortrait();
+						}
 					}
 				}
 			}
@@ -112,7 +127,7 @@ export function o13pcActorMixin(base) {
 			}
 		}
 		
-		_onAROverrideChange(adddiff, remdiff) {
+		_onAEOverrideChange(adddiff, remdiff) {
 			if (adddiff.system?.hasOwnProperty("maxwounds") || remdiff.system?.hasOwnProperty("maxwounds")) {
 				this.updateMaxWounds();
 			}
@@ -164,6 +179,10 @@ export function o13pcActorMixin(base) {
 		//Story
 		get storyActor() {
 			return [...game.actors].find(actor => actor.isStory && actor.hasPC(this))
+		}
+		
+		get siblingCharacters() {
+			return this.storyActor?.pcActors?.filter(actor => actor != this) ?? [];
 		}
 		
 		//Acts
@@ -234,6 +253,8 @@ export function o13pcActorMixin(base) {
 		
 		async setOwnArchetype(archetype) {
 			if (archetype.isArchetype) {
+				const shouldSyncImage = this.shouldSynchArchetypePortrait;
+				
 				await this.removeOwnArchetype();
 				
 				await this.createEmbeddedDocuments("Item", [archetype.toObject()]);
@@ -252,6 +273,7 @@ export function o13pcActorMixin(base) {
 				else {
 					await this.updateArchetypeItems();
 					await this.synctoArchetypeBackground();
+					if (shouldSyncImage) await this.synchArchetypePortrait();
 				}
 			}
 		}
@@ -259,9 +281,12 @@ export function o13pcActorMixin(base) {
 		async removeOwnArchetype() {
 			const oldArchetypes = [...this.items].filter(item => item.isArchetype);
 				
+			const shouldSyncImage = this.shouldSynchArchetypePortrait;
+				
 			await this.deleteEmbeddedDocuments("Item", oldArchetypes.map(archetype => archetype.id));
 			
 			await this.updateArchetypeItems();
+			if (shouldSyncImage) await this.synchArchetypePortrait();
 		}
 		
 		get ownArchetype() {
@@ -284,6 +309,27 @@ export function o13pcActorMixin(base) {
 			if (archetype) {
 				return this.createEmbeddedDocuments("Item", [...Object.values(archetype.perksData), ...Object.values(archetype.guaranteedGear)]);
 			}
+		}
+		
+		async synchArchetypePortrait() {
+			const archetype = this.archetype;
+			
+			if (archetype?.img) {
+				if (this.img != archetype.img) {
+					return this.update({img : archetype.img})
+				}
+			}
+			else {
+				return this.update({img : this.defaultPortrait})
+			}
+		}
+		
+		get hasArchetypePortrait() {
+			return this.img == this.archetype?.img;
+		}
+		
+		get shouldSynchArchetypePortrait() {
+			return this.hasArchetypePortrait || this.hasDefaultPortrait;
 		}
 		
 		get archetypePrepState() {
@@ -330,6 +376,12 @@ export function o13pcActorMixin(base) {
 		usePerk(id) {
 			if (this.hasPickedPerk(id)) {
 				this.pickedPerks[id]?.use();
+			}
+		}
+		
+		restorePerkUse(id) {
+			if (this.hasPickedPerk(id)) {
+				this.pickedPerks[id]?.restoreUse();
 			}
 		}
 		
@@ -400,7 +452,12 @@ export function o13pcActorMixin(base) {
 				const gearData = this.archetype?.unguaranteedGear[originid];
 
 				if (gearData) {
-					this.createEmbeddedDocuments("Item", [gearData]);
+					const data = foundry.utils.deepClone(gearData);
+					
+					data.system = data.system ?? {};
+					data.system.selectedinact = this.activeAct;
+					
+					this.createEmbeddedDocuments("Item", [data]);
 				}
 			}
 		}
@@ -415,6 +472,31 @@ export function o13pcActorMixin(base) {
 			if (selectedGearLength == this.selectableGearCount) return "ready";
 			
 			if (selectedGearLength > this.selectableGearCount) return "problem";
+		}
+		
+		getPostSelectedGear(act = undefined) {
+			let postSelectGear = this.inventory.filter(gear => gear.selectedinAct > 0);
+			
+			if (act != undefined) {
+				postSelectGear = postSelectGear.filter(gear => gear.selectedinAct == act);
+			}
+			
+			return postSelectGear ?? [];
+		}
+		
+		get canPostSelectGear() {
+			const perStory = this.system.postselectgearcount?.perstory ?? 0;
+			const perAct = this.system.postselectgearcount?.peract ?? 0;
+			
+			if (!perStory && !perAct) return false;
+			
+			const inStory = this.getPostSelectedGear();
+			
+			if (inStory.length < perStory) return true;
+			
+			const inAct = this.getPostSelectedGear(this.activeAct);
+			
+			if (inAct.length < perAct) return true;
 		}
 		
 		//Aspects
@@ -487,19 +569,63 @@ export function o13pcActorMixin(base) {
 			return !isNaN(this.getAspectData(aspect,true)?.targetNumber);
 		}
 		
+		get activeRollConfig() {
+			return this._activerollconfig;
+		}
+		
+		set activeRollConfig(rollConfig) {
+			if (rollConfig instanceof o13rollConfig) {
+				if (this._activerollconfig != rollConfig) this.closeActiveRollConfig();
+				
+				this._activerollconfig = rollConfig;
+			}
+		}
+		
+		getRollConfigData(aspectName, options = {}) {
+			const aspectData = this.getAspectData(aspectName, true);
+			const aspectModifiers = this.getAspectRollModifiers(aspectName);
+			
+			if (aspectData) {
+				const config = utils.combineRollOptions([aspectData, utils.rollOptionsFromModifiers(aspectModifiers), utils.expandRollData(options)]);
+
+				return config;
+			}
+		}
+		
+		async closeActiveRollConfig() {
+			if (this._activerollconfig) {
+				this._lastrolloptions = null;
+				return this._activerollconfig.close();
+			}
+			return false;
+		}
+		
 		async rollAspect(aspectName, options = {}, quickRoll = false) {
 			if (this.canRollAspect(aspectName)) {
-				const aspectData = this.getAspectData(aspectName, true);
-				const aspectModifiers = this.getAspectRollModifiers(aspectName);
+				const config = this.getRollConfigData(aspectName, options);
 
-				if (aspectData) {
-					const config = utils.combineRollOptions([aspectData, utils.rollOptionsFromModifiers(aspectModifiers), utils.expandRollData(options)]);
+				if (config) {
+					await this.closeActiveRollConfig();
 
-					new o13rollConfig(this, {...config, aspect : aspectName}, quickRoll);
+					this._lastrolloptions = options;
+					this._activerollconfig = new o13rollConfig(this, {...config, aspect : aspectName}, quickRoll);
+					
+					return this._activerollconfig.wait();
 				}
 			}
 			else {
 				ui.notifications.warn(game.i18n.localize("13omens.warnings.selectRating"), {console : false});
+			}
+		}
+		
+		async updateActiveRollConfig(options = {}) {
+			if (this._activerollconfig) {
+				this._lastrolloptions = {...this._lastrolloptions, ...options};
+				
+				const config = this.getRollConfigData(this._activerollconfig.aspect, this._lastrolloptions);
+				config.aspect = this._activerollconfig.aspect;
+
+				return this._activerollconfig.updateData(config);
 			}
 		}
 		
@@ -801,6 +927,11 @@ export function o13pcActorMixin(base) {
 			this.system._activeact = this.activeAct;
 								
 			this.system.selectablegearcount = this.archetype?.selectableGearCount ?? 0;
+			
+			this.system.postselectgearcount = {
+				perstory: 0,
+				peract: 0
+			};
 				
 			this.system.cheatdeathamount = { //these give the maximum
 				perstory: 1,